BaseModel renamed to PErmutectModel (probably eventually rename ArtifactModel once full absorption)
